Rename operator component to ops
