--- cmd/operator/OWNERS
@@ -1,2 +1,0 @@
-labels:
-  - area/operator
