+++ cmd/ops/OWNERS
@@ -1,0 +1,2 @@
+labels:
+  - area/ops
