+++ pkg/app/ops/handler/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
         "handler.go",
         ":templates.embed",  #keep
     ],
-    importpath = "github.com/pipe-cd/pipe/pkg/app/operator/handler",
+    importpath = "github.com/pipe-cd/pipe/pkg/app/ops/handler",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/config:go_default_library",
