+++ pkg/app/ops/insightcollector/BUILD.bazel
@@ -2,7 +2,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["cleaner.go"],
-    importpath = "github.com/pipe-cd/pipe/pkg/app/operator/modelcleaner",
+    srcs = ["collector.go"],
+    importpath = "github.com/pipe-cd/pipe/pkg/app/ops/insightcollector",
     visibility = ["//visibility:public"],
 )
