+++ pkg/app/ops/insightcollector/collector.go
